requirements_tray.txt
pystray==0.19.2
Pillow==10.0.0
psutil==5.9.5